10-20 修改了数据库 重新生成

+++ src/main/java/com/code/pojo/AnswerRecord.java
@@ -1,14 +1,14 @@
 package com.code.pojo;
 
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.Version;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableField;
+import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
-
-import java.io.Serializable;
 
 /**
  * <p>
@@ -16,7 +16,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author xiaoshuai
- * @since 2022-10-17
+ * @since 2022-10-20
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -27,9 +27,9 @@ public class AnswerRecord implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "sutdent_item_id", type = IdType.AUTO)
-    private Long sutdentItemId;
+    private Integer sutdentItemId;
 
-    private Long itemsetId;
+    private Integer itemsetId;
 
     @TableField("checkingRate")
     private Float checkingRate;
