+++ src/main/java/com/code/pojo/ItemItemset.java
@@ -1,13 +1,13 @@
 package com.code.pojo;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.Version;
 import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableField;
+import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
-
-import java.io.Serializable;
 
 /**
  * <p>
@@ -15,7 +15,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author xiaoshuai
- * @since 2022-10-17
+ * @since 2022-10-20
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -25,11 +25,11 @@ public class ItemItemset implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "item_id", type = IdType.AUTO)
-    private Long itemId;
+    private Integer itemId;
 
-    private Long itemsetId;
+    private Integer itemsetId;
 
-    private Long sortId;
+    private Integer sortId;
 
     @TableField("passingRate")
     private Float passingRate;
