+++ src/main/java/com/code/pojo/Testcase.java
@@ -1,12 +1,12 @@
 package com.code.pojo;
 
 import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.Version;
 import com.baomidou.mybatisplus.annotation.TableId;
+import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
-
-import java.io.Serializable;
 
 /**
  * <p>
@@ -14,7 +14,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author xiaoshuai
- * @since 2022-10-17
+ * @since 2022-10-20
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -24,9 +24,9 @@ public class Testcase implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "testcase_id", type = IdType.AUTO)
-    private Long testcaseId;
+    private Integer testcaseId;
 
-    private Long itemId;
+    private Integer itemId;
 
     private String testcaseInput;
